Lecture 12 Completed

+++ Lecture 12/utils/cloudinary.js
@@ -4,21 +4,19 @@ import { v2 as cloudinary } from "cloudinary";
 cloudinary.config({
 	cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
 	api_key: process.env.CLOUDINARY_API_KEY,
-	api_secret: process.env.CLOUDINARY_API_SECRETS,
+	api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
 export const fileupload = async (filepath) => {
 	try {
 		if (!filepath) return null;
-		const response = await cloudinary.v2.uploader
-			.upload(filepath, {
-				resource_type: "video",
-				overwrite: true,
-			})
-			.then((result) => console.log(result));
+		const response = await cloudinary.uploader.upload(filepath, {
+			resource_type: "auto",
+		});
+		fs.unlinkSync(filepath);
 		return response;
 	} catch (error) {
-		fs.unlinkSync(filepath);
+		if (fs.existsSync(filepath)) fs.unlinkSync(filepath);
 		console.log(error);
 		return null;
 	}
